requirements.txt
# Requires: Python >= 3.10
accelerate==1.8.1
clip==0.2.0
datasets>=2.21,<3.0
diffusers==0.34.0
fairscale==0.4.13
lpips>=0.1.4,<0.2
matplotlib>=3.10,<3.11
numpy>=1.26,<2.0
opencv-python>=4.11.0.86,<5.0
pillow>=10.3,<11.0
protobuf>=5.27
pytorch-fid>=0.3.0,<0.4
safetensors>=0.4.4,<0.5
scikit-image>=0.25,<0.26
scipy>=1.14,<1.15
sentencepiece==0.2.1
timm==1.0.16
torch>=2.6,<2.7
torchvision>=0.21,<0.22
torchmetrics==1.7.3
transformers==4.42.4
tqdm>=4.66,<5.0
zstandard>=0.23,<0.24
